+++ sdk/communication/azure-communication-common/src/test/java/com/azure/android/communication/common/CommunicationTokenCredentialTest.java
@@ -3,8 +3,6 @@
 
 package com.azure.android.communication.common;
 
-import com.azure.android.core.credential.AccessToken;
-
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -55,7 +53,7 @@ public class CommunicationTokenCredentialTest {
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
         countDownLatch.await();
 
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(refreshedToken, accessToken.getToken());
         assertEquals(1, mockTokenRefresher.getCallCount());
@@ -80,7 +78,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
 
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true, tokenString));
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(tokenString, accessToken.getToken());
         assertEquals(0, mockTokenRefresher.getCallCount());
@@ -124,7 +122,7 @@ public class CommunicationTokenCredentialTest {
         String tokenString = TokenStubHelper.createTokenString(expiryEpochSecond);
 
         CommunicationTokenCredential credential = new CommunicationTokenCredential(tokenString);
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(tokenString, accessToken.getToken());
         assertEquals(expiryEpochSecond, accessToken.getExpiresAt().toEpochSecond());
@@ -137,7 +135,7 @@ public class CommunicationTokenCredentialTest {
         String tokenString = TokenStubHelper.createTokenString(expiryEpochSecond);
 
         CommunicationTokenCredential credential = new CommunicationTokenCredential(tokenString);
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(tokenString, accessToken.getToken());
         assertEquals(expiryEpochSecond, accessToken.getExpiresAt().toEpochSecond());
@@ -150,7 +148,7 @@ public class CommunicationTokenCredentialTest {
         CommunicationTokenCredential credential = new CommunicationTokenCredential(tokenString);
 
         credential.dispose();
-        Future<AccessToken> accessTokenFuture = credential.getToken();
+        Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
 
         assertTrue(accessTokenFuture.isDone());
         assertTrue(accessTokenFuture.isCancelled());
@@ -165,7 +163,7 @@ public class CommunicationTokenCredentialTest {
         mockTokenRefresher.setToken(refreshedToken);
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false));
 
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(refreshedToken, accessToken.getToken());
         assertEquals(1, mockTokenRefresher.getCallCount());
@@ -201,17 +199,17 @@ public class CommunicationTokenCredentialTest {
         // Set up blocked multithreaded calls
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
         int numCalls = 5;
-        List<Future<AccessToken>> accessTokenFutures = new ArrayList<>();
+        List<Future<CommunicationAccessToken>> accessTokenFutures = new ArrayList<>();
         ExecutorService executorService = Executors.newFixedThreadPool(numCalls);
         for (int i = 0; i < numCalls; i++) {
-            Future<AccessToken> accessTokenFuture = executorService.submit(() -> credential.getToken().get());
+            Future<CommunicationAccessToken> accessTokenFuture = executorService.submit(() -> credential.getToken().get());
             accessTokenFutures.add(accessTokenFuture);
         }
 
         // Unblock refresh and wait for results
         blockedRefresh.run();
-        Set<AccessToken> accessTokenResults = new HashSet<>();
-        for (Future<AccessToken> accessTokenFuture : accessTokenFutures) {
+        Set<CommunicationAccessToken> accessTokenResults = new HashSet<>();
+        for (Future<CommunicationAccessToken> accessTokenFuture : accessTokenFutures) {
             accessTokenResults.add(accessTokenFuture.get());
         }
         executorService.shutdown();
@@ -226,7 +224,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(tokenString, accessToken.getToken());
         assertEquals(0, mockTokenRefresher.getCallCount());
@@ -241,17 +239,17 @@ public class CommunicationTokenCredentialTest {
         // Set up blocked multithreaded calls
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
         int numCalls = 5;
-        List<Future<AccessToken>> accessTokenFutures = new ArrayList<>();
+        List<Future<CommunicationAccessToken>> accessTokenFutures = new ArrayList<>();
         ExecutorService executorService = Executors.newFixedThreadPool(numCalls);
         for (int i = 0; i < numCalls; i++) {
-            Future<AccessToken> accessTokenFuture = executorService.submit(() -> credential.getToken().get());
+            Future<CommunicationAccessToken> accessTokenFuture = executorService.submit(() -> credential.getToken().get());
             accessTokenFutures.add(accessTokenFuture);
         }
 
         // Unblock refresh and wait for results
         blockedRefresh.run();
-        Set<AccessToken> accessTokenResults = new HashSet<>();
-        for (Future<AccessToken> accessTokenFuture : accessTokenFutures) {
+        Set<CommunicationAccessToken> accessTokenResults = new HashSet<>();
+        for (Future<CommunicationAccessToken> accessTokenFuture : accessTokenFutures) {
             accessTokenResults.add(accessTokenFuture.get());
         }
         executorService.shutdown();
@@ -268,7 +266,7 @@ public class CommunicationTokenCredentialTest {
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(300);
         mockTokenRefresher.setToken(refreshedToken);
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
-        AccessToken accessToken = credential.getToken().get();
+        CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertNotEquals(tokenString, accessToken.getToken());
         assertEquals(refreshedToken, accessToken.getToken());
@@ -301,10 +299,10 @@ public class CommunicationTokenCredentialTest {
         mockTokenRefresher.setToken(refreshedToken);
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
-        Set<AccessToken> accessTokens = new HashSet<>();
+        Set<CommunicationAccessToken> accessTokens = new HashSet<>();
         int numCalls = 3;
         for (int i = 0; i < numCalls; i++) {
-            AccessToken accessToken = credential.getToken().get();
+            CommunicationAccessToken accessToken = credential.getToken().get();
             accessTokens.add(accessToken);
         }
 
@@ -325,17 +323,17 @@ public class CommunicationTokenCredentialTest {
         // Set up blocked multithreaded calls
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
         int numCalls = 5;
-        List<Future<AccessToken>> accessTokenFutures = new ArrayList<>();
+        List<Future<CommunicationAccessToken>> accessTokenFutures = new ArrayList<>();
         ExecutorService executorService = Executors.newFixedThreadPool(numCalls);
         for (int i = 0; i < numCalls; i++) {
-            Future<AccessToken> accessTokenFuture = executorService.submit(() -> credential.getToken().get());
+            Future<CommunicationAccessToken> accessTokenFuture = executorService.submit(() -> credential.getToken().get());
             accessTokenFutures.add(accessTokenFuture);
         }
 
         // Unblock refresh and wait for results
         blockedRefresh.run();
-        Set<AccessToken> accessTokenResults = new HashSet<>();
-        for (Future<AccessToken> accessTokenFuture : accessTokenFutures) {
+        Set<CommunicationAccessToken> accessTokenResults = new HashSet<>();
+        for (Future<CommunicationAccessToken> accessTokenFuture : accessTokenFutures) {
             accessTokenResults.add(accessTokenFuture.get());
         }
         executorService.shutdown();
@@ -352,7 +350,7 @@ public class CommunicationTokenCredentialTest {
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false));
 
         credential.dispose();
-        Future<AccessToken> accessTokenFuture = credential.getToken();
+        Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
 
         assertTrue(accessTokenFuture.isDone());
         assertTrue(accessTokenFuture.isCancelled());
@@ -368,9 +366,9 @@ public class CommunicationTokenCredentialTest {
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
 
-        Future<AccessToken> accessTokenFuture = credential.getToken();
+        Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
         blockedRefresh.run();
-        AccessToken accessToken = accessTokenFuture.get();
+        CommunicationAccessToken accessToken = accessTokenFuture.get();
 
         assertEquals(refreshedToken, accessToken.getToken());
         assertEquals(1, mockTokenRefresher.getCallCount());
@@ -386,7 +384,7 @@ public class CommunicationTokenCredentialTest {
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
 
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
-        Future<AccessToken> accessTokenFuture = credential.getToken();
+        Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
         credential.dispose();
         blockedRefresh.run();
 
@@ -405,7 +403,7 @@ public class CommunicationTokenCredentialTest {
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
 
         CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
-        Future<AccessToken> accessTokenFuture = credential.getToken();
+        Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
         credential.dispose();
         blockedRefresh.run();